Put all data through the stream
It seems obvious in retrospect, but defining extra properties and events
on stream objects is not a good idea, since it makes them incomposible.
In particular, nothing is happening right now because streams are never
seen as "warm".

So instead of communicating liveness "out of band", switch to just
sending a simple `{nowLive: true}` object once a metric stream is warm.