Mark workerTypes with 0 pending tasks as not pending
Although we can't introspect a nonzero queue, a queue with 0 items in it
means that the current pending time for that queue is zero.  We only
poll this at startup, when other information about that workerType is
not available, so that we can start getting accurate information more
quickly without overwhelming the queue with API calls.